pyproject.toml
[project]
name = "telebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg==0.29.0",
    "nest-asyncio>=1.6.0",
    "psycopg[binary]>=3.2.6",
    "python-dotenv==1.0.1",
    "python-telegram-bot[asyncio]==20.8",
    "pytz>=2025.2",
]
